cron fixes

+++ RecordData/uploader.py
@@ -8,7 +8,8 @@ from oauth2client.service_account import ServiceAccountCredentials
 from httplib2 import Http
 import os,shutil
 
-import py7zr
+#import py7zr
+import tarfile
 import datetime
 
 gauth = GoogleAuth()
@@ -34,18 +35,18 @@ def make_ping():
 
 def upload_youtube_data():  
     todaydate = datetime.datetime.today().strftime('%Y-%m-%d')
-    zip_filename = '{}.7z'.format(todaydate)
+    zip_filename = '{}.tar.gz'.format(todaydate)
     
     # Create an archive
-    with py7zr.SevenZipFile(zip_filename, 'w') as archive:
-        archive.write('../cron_scraper.log')
+    with tarfile.open(zip_filename, 'w:gz') as archive:
+        archive.add('../cron_scraper.log')
         try:
-            archive.write('../cron_updater.log')
+            archive.add('../cron_updater.log')
         except:
             pass
-        archive.write('errors.log')
-        archive.writeall('comments/')
-        archive.writeall('viewers/')
+        archive.add('errors.log')
+        archive.add('comments/')
+        archive.add('viewers/')
     
     # Upload archive
     file1 = drive.CreateFile(
@@ -53,24 +54,52 @@ def upload_youtube_data():
                               'parents':[{'id':'1A_AtZGmB8S34HB4TC1URyjT9HHvOn_o2'}]
                               }
             )
+    file1.SetContentFile(zip_filename)
     file1.Upload()
     
-    # Remove all old files
-    folders = ['viewers','comments']
-    for folder in folders:
-        for filename in os.listdir(folder):
-            file_path = os.path.join(folder,filename)
-            try:
-                if os.path.isfile(file_path) or os.path.islink(file_path):
-                    os.unlink(file_path)
-                elif os.path.isdir(file_path):
-                    shutil.rmtree(file_path)
-            except Exception as e:
-                print('Failed to delete %s. Reason: %s' % (file_path,e))
-    # Remove archive
-    if os.path.exists(zip_filename):
-        os.remove(zip_filename)
-    else:
-        print('Cannot delete archive {}, doesn`t exist'.format(zip_filename))
+    remove_files = False
     
+    if remove_files:
+        # Remove all old files
+        folders = ['viewers','comments']
+        for folder in folders:
+            for filename in os.listdir(folder):
+                file_path = os.path.join(folder,filename)
+                try:
+                    if os.path.isfile(file_path) or os.path.islink(file_path):
+                        os.unlink(file_path)
+                    elif os.path.isdir(file_path):
+                        shutil.rmtree(file_path)
+                except Exception as e:
+                    print('Failed to delete %s. Reason: %s' % (file_path,e))
+        # Remove archive
+        if os.path.exists(zip_filename):
+            os.remove(zip_filename)
+        else:
+            print('Cannot delete archive {}, doesn`t exist'.format(zip_filename))
+
+def share_error(error):  
+    # Upload archive
+    file1 = drive.CreateFile(
+            {'title':error,
+                              'parents':[{'id':'1Us3pMvn7H6Bga9WcK_NyMqkS0jhy9zI5'}]
+                              }
+            )
+    file1.SetContentFile(error)
+    file1.Upload()
+
+def test():
+    todaydate = datetime.datetime.today().strftime('%Y-%m-%d')
+    zip_filename = '{}.tar.gz'.format(todaydate)
+    with tarfile.open(zip_filename, 'w:gz') as archive:
+        archive.add('../cron_scraper.log')
+        try:
+            archive.add('../cron_updater.log')
+        except:
+            pass
+        archive.add('errors.log')
+        archive.add('comments/')
+        archive.add('viewers/')
+#test()
+
 #upload_youtube_data()
